create component status and reuse every where is needed
